all json visible on front end through json except invoices and invoice-items

+++ src/main/java/zero/repository/ProductRepository.java
@@ -8,6 +8,9 @@ import zero.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    //List<Product> findByNameContainingIgnoreCase(String name);      //make method maybe?
 }
